fix: web refactor case error
