+++ packages/mock/src/utils.ts
@@ -1,5 +1,6 @@
 import { BootstrapStarter } from '@midwayjs/bootstrap';
 import {
+  ContainerLoader,
   IMidwayApplication,
   IMidwayFramework,
   MidwayFrameworkType,
@@ -32,6 +33,7 @@ export async function create<
 ): Promise<T> {
   process.env.MIDWAY_TS_MODE = 'true';
   clearAllModule();
+  ContainerLoader.parentApplicationContext = null;
   let framework: T = null;
   let DefaultFramework = null;
