+++ packages/web/app/extend/application.js
@@ -1,0 +1,17 @@
+module.exports = {
+  get appOptions() {
+    return this.options;
+  },
+
+  get midwayWebFramework() {
+    return this.appOptions['webFramework'];
+  },
+
+  get applicationContext() {
+    return this.midwayWebFramework.getApplicationContext();
+  },
+
+  get baseDir() {
+    return this.loader.baseDir;
+  },
+};
