+++ packages/web/src/application.ts
@@ -1,9 +1,8 @@
-import type { MidwayWebFramework } from './framework';
-import { RouterParamValue } from '@midwayjs/decorator';
+// import type { MidwayWebFramework } from './framework';
+// import { RouterParamValue } from '@midwayjs/decorator';
 import { parseNormalDir } from './utils';
 import * as extend from 'extend2';
 import { EggAppInfo } from 'egg';
-import { IMidwayWebApplication } from './interface';
 import { join } from 'path';
 
 const {
@@ -18,24 +17,17 @@ const EGG_PATH = Symbol.for('egg#eggPath');
 
 export const createAppWorkerLoader = AppWorkerLoader => {
   class EggAppWorkerLoader extends (AppWorkerLoader as any) {
-    app: IMidwayWebApplication & {
-      appOptions: {
-        typescript?: boolean;
-        isTsMode?: boolean;
-      };
-      appDir: string;
-      baseDir: string;
-    };
+    app: any;
 
     getEggPaths() {
       if (!this.appDir) {
         // 这里的逻辑是为了兼容老 cluster 模式
-        if (this.app.appOptions.typescript || this.app.appOptions.isTsMode) {
+        if (this.app.options.typescript || this.app.options.isTsMode) {
           process.env.EGG_TYPESCRIPT = 'true';
         }
         const result = parseNormalDir(
-          this.app.appOptions['baseDir'],
-          this.app.appOptions.isTsMode
+          this.app.options['baseDir'],
+          this.app.options.isTsMode
         );
         this.baseDir = result.baseDir;
         this.options.baseDir = this.baseDir;
@@ -72,12 +64,12 @@ export const createAgentWorkerLoader = AppWorkerLoader => {
   class EggAppWorkerLoader extends (AppWorkerLoader as any) {
     getEggPaths() {
       if (!this.appDir) {
-        if (this.app.appOptions.typescript || this.app.appOptions.isTsMode) {
+        if (this.app.options.typescript || this.app.options.isTsMode) {
           process.env.EGG_TYPESCRIPT = 'true';
         }
         const result = parseNormalDir(
-          this.app.appOptions['baseDir'],
-          this.app.appOptions.isTsMode
+          this.app.options['baseDir'],
+          this.app.options.isTsMode
         );
         this.baseDir = result.baseDir;
         this.options.baseDir = this.baseDir;
@@ -123,42 +115,42 @@ export const createEggApplication = Application => {
     get [EGG_PATH]() {
       return __dirname;
     }
-
-    get appOptions() {
-      return this.options;
-    }
-
-    get midwayWebFramework(): MidwayWebFramework {
-      return this.appOptions['webFramework'];
-    }
-
-    get applicationContext() {
-      return this.midwayWebFramework.getApplicationContext();
-    }
-
-    getApplicationContext() {
-      return this.applicationContext;
-    }
-
-    generateController(
-      controllerMapping: string,
-      routeArgsInfo?: RouterParamValue[],
-      routerResponseData?: any[]
-    ) {
-      return this.midwayWebFramework.generateController(
-        controllerMapping,
-        routeArgsInfo,
-        routerResponseData
-      );
-    }
-
-    async generateMiddleware(middlewareId: string) {
-      return this.midwayWebFramework.generateMiddleware(middlewareId);
-    }
-
-    get baseDir() {
-      return this.loader.baseDir;
-    }
+    //
+    // get appOptions() {
+    //   return this.options;
+    // }
+    //
+    // get midwayWebFramework(): MidwayWebFramework {
+    //   return this.appOptions['webFramework'];
+    // }
+    //
+    // get applicationContext() {
+    //   return this.midwayWebFramework.getApplicationContext();
+    // }
+    //
+    // getApplicationContext() {
+    //   return this.applicationContext;
+    // }
+    //
+    // generateController(
+    //   controllerMapping: string,
+    //   routeArgsInfo?: RouterParamValue[],
+    //   routerResponseData?: any[]
+    // ) {
+    //   return this.midwayWebFramework.generateController(
+    //     controllerMapping,
+    //     routeArgsInfo,
+    //     routerResponseData
+    //   );
+    // }
+    //
+    // async generateMiddleware(middlewareId: string) {
+    //   return this.midwayWebFramework.generateMiddleware(middlewareId);
+    // }
+    //
+    // get baseDir() {
+    //   return this.loader.baseDir;
+    // }
   }
 
   return EggApplication as any;
@@ -177,26 +169,6 @@ export const createEggAgent = Agent => {
 
     get [EGG_PATH]() {
       return __dirname;
-    }
-
-    get appOptions() {
-      return this.options;
-    }
-
-    get midwayWebFramework(): MidwayWebFramework {
-      return this.appOptions['webFramework'];
-    }
-
-    get applicationContext() {
-      return this.midwayWebFramework.getApplicationContext();
-    }
-
-    getApplicationContext() {
-      return this.applicationContext;
-    }
-
-    get baseDir() {
-      return this.loader.baseDir;
     }
   }
