+++ packages/web/test/fixtures/feature/base-app-middleware/src/middleware/global.ts
@@ -5,7 +5,6 @@ import { IMidwayWebContext } from '../../../../../../src';
 export class GlobalMiddleware1 {
   resolve() {
     return async (ctx: IMidwayWebContext, next) => {
-      console.log('1///');
       ctx.state.a = '1111';
       await next();
     }
